Refactor cds taric sync worker
